+++ tests/TypeTest.php
@@ -15,13 +15,13 @@ final class TypeTest extends TestCase
     #[Test]
     public function it_will_return_null_if_the_type_is_no_supported(): void
     {
-        self::assertNull(Type::tryFromValue([]));
+        self::assertNull(Type::tryFromVariable([]));
     }
 
     #[Test]
     public function it_will_return_true_if_the_type_is_valid_but_its_value_is_invalid(): void
     {
-        self::assertTrue(Type::Integer->equals(Type::tryFromValue(1_000_000_000_000_000)));
+        self::assertTrue(Type::Integer->equals(Type::tryFromVariable(1_000_000_000_000_000)));
     }
 
     #[Test]
@@ -35,8 +35,8 @@ final class TypeTest extends TestCase
     #[DataProvider('itemTypeProvider')]
     public function it_can_tell_the_item_type(mixed $value, Type $expectedType): void
     {
-        self::assertTrue($expectedType->equals(Type::fromValue($value)));
-        self::assertTrue($expectedType->equals(Type::tryFromValue($value)));
+        self::assertTrue($expectedType->equals(Type::fromVariable($value)));
+        self::assertTrue($expectedType->equals(Type::tryFromVariable($value)));
     }
 
     /**